simple sort number

+++ src/main/java/fibonacci_number/FibonacciMain.java
@@ -5,13 +5,12 @@ import java.util.Scanner;
 public class FibonacciMain {
     public static void main(String[] args) {
         System.out.println("Insert Fibonnaci number series");
-
-        int fibseries = new Scanner(System.in).nextInt();
-        int secondNumber = 1;
-        int firstNumber = 0;
+        long fibseries = new Scanner(System.in).nextInt();
+        long secondNumber = 1;
+        long firstNumber = 0;
         for (; fibseries > 1; fibseries--) {
             if (firstNumber == 0) System.out.print(1+", ");
-            int fibNumber = firstNumber+secondNumber;
+            long fibNumber = firstNumber+secondNumber;
             firstNumber = secondNumber;
             secondNumber = fibNumber;
             if (fibseries!=2){
